% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blink.R
\name{blink}
\alias{blink}
\title{A Blinking Function}
\usage{
blink(times = 3, color = "blue", printcomm = FALSE)
}
\arguments{
\item{times}{the number of times to blink}

\item{Color}{the name of the color to blink

The value for Color may be one of the following:
white,red,lime,blue,yellow,cyan,magenta,silver,gray,maroon,olive,green,purple,teal or navy.

The times variable is expecting an int greater than zero.}
}
\description{
This function provides a limited wrapper to the Blink1 device. To learn more about 
the device https://blink1.thingm.com/
}
\examples{
blink()


}
\keyword{package}

