% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-cmdstan.R
\name{build_cmdstan}
\alias{build_cmdstan}
\title{Build cmdstan with optimizations}
\usage{
build_cmdstan()
}
\description{
From the Stan forums (\href{https://discourse.mc-stan.org/t/speedup-by-using-external-blas-lapack-with-cmdstan-and-cmdstanr-py/25441/41}{link}).
}
\examples{
\dontrun{
build_cmdstan()
}
}
