% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-utils.R
\name{check_all_diagnostics}
\alias{check_all_diagnostics}
\title{Check all Stan diagnostics}
\usage{
check_all_diagnostics(fit)
}
\arguments{
\item{fit}{A \code{\link[rstan:stanfit-class]{stanfit}} object.}
}
\value{
A named vector of results from the various checks.
}
\description{
Check all Stan diagnostics
}
\examples{
\dontrun{
checks <- check_all_diagnostics()
checks["divergences"]
}
}
