% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-utils.R
\name{check_divergences}
\alias{check_divergences}
\title{Check for transitions that ended with a divergence}
\usage{
check_divergences(fit)
}
\arguments{
\item{fit}{A \code{\link[rstan:stanfit-class]{stanfit}} object.}
}
\value{
The number of divergences in the fit.
}
\description{
Check for transitions that ended with a divergence
}
\examples{
\dontrun{
n_divs <- check_divergences(fit)
}
}
