% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-utils.R
\name{check_energy}
\alias{check_energy}
\title{Check the energy Bayesian fraction of missing information (E-BFMI)}
\usage{
check_energy(fit)
}
\arguments{
\item{fit}{A \code{\link[rstan:stanfit-class]{stanfit}} object.}
}
\value{
1 if there were E-BFMI warnings and 0 otherwise.
}
\description{
Check the energy Bayesian fraction of missing information (E-BFMI)
}
\examples{
\dontrun{
check_energy(fit)
}
}
