% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_for_packages}
\alias{check_for_packages}
\title{Check for installed packages}
\usage{
check_for_packages(packages)
}
\arguments{
\item{packages}{A character vector of one or more package names.}
}
\description{
Check for installed packages
}
\examples{
\dontrun{
check_for_namespace(c("data.table", "dplyr"))
}
}
