% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-utils.R
\name{check_n_eff}
\alias{check_n_eff}
\title{Checks the effective sample size per iteration}
\usage{
check_n_eff(fit)
}
\arguments{
\item{fit}{A \code{\link[rstan:stanfit-class]{stanfit}} object.}
}
\value{
1 if n_eff / iter is low and 0 otherwise.
}
\description{
Checks the effective sample size per iteration
}
\examples{
\dontrun{
check_n_eff(fit)
}
}
