% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-utils.R
\name{check_treedepth}
\alias{check_treedepth}
\title{Check transitions that ended prematurely due to maximum tree depth limit}
\usage{
check_treedepth(fit, max_depth = 10)
}
\arguments{
\item{fit}{A \code{\link[rstan:stanfit-class]{stanfit}} object.}

\item{max_depth}{The tree depth limit.}
}
\value{
The number of iterations saturating the tree depth limit.
}
\description{
Check transitions that ended prematurely due to maximum tree depth limit
}
\examples{
\dontrun{
check_treedepth(fit, 10)
}
}
