% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{col_order}
\alias{col_order}
\title{Example data for TT package}
\format{
character vector
}
\usage{
col_order
}
\description{
character vector to demonstrate ordering option. See example of \link{treetable}
}
\keyword{datasets}
