% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-package-tools.R
\name{extract_divergences}
\alias{extract_divergences}
\title{Extract divergences from a variety of Stan packages}
\usage{
extract_divergences(x, ...)
}
\arguments{
\item{x}{A Stan MCMC model fit.}

\item{...}{Arguments passed to S3 methods if applicable.}
}
\value{
A vector containing the number of divergences per iteration.
}
\description{
Extract divergences from a variety of Stan packages
}
\examples{
\dontrun{
fit <- brms::brm(y ~ x, data = datalist)
extract_divergences(fit)
}
}
