% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-package-tools.R
\name{extract_treedepths}
\alias{extract_treedepths}
\title{Extract tree depths from a variety of Stan packages}
\usage{
extract_treedepths(x, ...)
}
\arguments{
\item{x}{A Stan MCMC model fit.}

\item{...}{Arguments passed to S3 methods if applicable.}
}
\value{
A vector containing the tree depths per iteration.
}
\description{
Extract tree depths from a variety of Stan packages
}
\examples{
\dontrun{
fit <- brms::brm(y ~ x, data = datalist)
extract_treedepths(fit)
}
}
