% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadpkg.R
\name{loadpkg}
\alias{loadpkg}
\title{A Package Loading Function}
\usage{
loadpkg(x)
}
\arguments{
\item{x}{a character string that is a valid R package.}
}
\description{
This function uses `library' to load the package if installed and installs and then loads the package otherwise.
}
\examples{
loadpkg("knitr")
}
\keyword{package}

