% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{org}
\alias{org}
\title{Example data for TT package}
\format{
'data.tree' object with 7 attributes in each node, value of the leaf attributes is some random number, parents value is cumulative sum of children
The attributes are:\cr
\itemize{
\item name of the node\cr
\item six consecutive months of the year (\code{Jan-2021, Feb-2021....})\cr
}
}
\usage{
org
}
\description{
'data.tree' object that can be used as an example to see how table widget looks like.
Each node of 'data.tree' store attributes (dates) with some numbers.
}
\keyword{datasets}
