% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-utils.R
\name{partition_divergences}
\alias{partition_divergences}
\title{Separate parameter arrays into divergent and non-divergent transitions}
\usage{
partition_divergences(fit)
}
\arguments{
\item{fit}{A \code{\link[rstan:stanfit-class]{stanfit}} object.}
}
\value{
A list, separating divergent transitions from non-devergent.
}
\description{
Separate parameter arrays into divergent and non-divergent transitions
}
\examples{
\dontrun{
partition_divergences(fit)
}
}
