% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residout.R
\name{residout}
\alias{residout}
\title{A Residuals Function}
\usage{
residout(x, y)
}
\arguments{
\item{y}{response vector}

\item{X}{a design matrix y response vector}
}
\description{
This function allows you to output a residuals table in latex format.
}
\examples{
residout()
}
\keyword{residuals}

