% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setdiff3.R
\name{setdiff3}
\alias{setdiff3}
\title{Three-way \code{\link[=setdiff]{setdiff()}} of two collections}
\usage{
setdiff3(a, b)
}
\arguments{
\item{a}{A collection of R objects.}

\item{b}{A collection of R objects.}
}
\value{
A list with objects only in \code{a}, only in \code{b},
and in both collections.
}
\description{
Three-way \code{\link[=setdiff]{setdiff()}} of two collections
}
