% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_na.R
\name{table_na}
\alias{table_na}
\title{Cross tabulation with NAs included by default}
\usage{
table_na(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[=table]{table()}}.}
}
\value{
A \code{\link[=table]{table()}} object.
}
\description{
Cross tabulation with NAs included by default
}
