% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-tmth.R
\name{theme_tmth}
\alias{theme_tmth}
\title{\emph{Trees, maps, and theorems} plot theme}
\usage{
theme_tmth(
  base_size = 12,
  base_family = "Source Sans Pro",
  base_line_size = 0.4,
  base_rect_size = 0.4,
  aspect_ratio = 2/(1 + sqrt(5))
)
}
\arguments{
\item{base_size}{Foundational font size}

\item{base_family}{Foundational font family}

\item{base_line_size}{Foundational line size}

\item{base_rect_size}{Foundational rectangle size}

\item{aspect_ratio}{Foundational aspect ratio (default: golden mean)}
}
\value{
A new theme derived from \code{\link[ggplot2:ggtheme]{ggplot2::theme_bw()}}
}
\description{
A ggplot2 theme following some guidelines from the book
\emph{Trees, maps, and theorems}.
}
\examples{
\dontrun{
ggplot(mtcars) +
  geom_point(aes(wt, mpg)) +
  scale_x_continuous(limits = c(1, 6), breaks = seq(1, 6)) +
  scale_y_continuous(limits = c(10, 35)) +
  theme_tmth() +
  coord_capped_cart(left = "both", bottom = "both")
}
}
