% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.R
\name{trim}
\alias{trim}
\title{A Trim Function}
\usage{
trim(x)
}
\arguments{
\item{x}{a character string}
}
\description{
This function trims the whitespace on both sides of a string.
}
\examples{
trim("  Hello World ")
}
\keyword{trim}

