% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts-pdf-document.R
\name{ts_pdf_document}
\alias{ts_pdf_document}
\title{Generate a PDF document with custom options}
\usage{
ts_pdf_document(...)
}
\arguments{
\item{...}{Arguments passed to \link[bookdown:html_document2]{bookdown::pdf_document2}}
}
\description{
TODO: Create our own preprocessor to avoid setting so many options
in the preamble.
}
