% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tshfr-package.R
\docType{package}
\name{tshfr-package}
\alias{tshfr}
\alias{tshfr-package}
\title{tshfr: R Tools for Data Science Work}
\description{
A collection of R tools for data science and other analysis.
}
\author{
\strong{Maintainer}: Tom Shafer \email{tom@tshafer.com}

}
\keyword{internal}
