% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tt-package.R
\docType{package}
\name{tt-package}
\alias{tt}
\alias{tt-package}
\title{tt: Personal R Tools}
\description{
A collection of R tools for data science and other analysis.
}
\author{
\strong{Maintainer}: Tom Shafer \email{tom@tshafer.com}

}
\keyword{internal}
