% Generated by roxygen2 (4.0.1): do not edit by hand
\name{residout}
\alias{residout}
\title{A Residuals Function}
\usage{
residout(x, y)
}
\arguments{
\item{X}{a design matrix y response vector}

\item{y}{response vector}
}
\description{
This function allows you to output a residuals table in latex format.
}
\examples{
residout()
}
\keyword{residuals}

